# apps/web/public/api-spec.yaml
openapi: "3.1.0"
info:
  title: Vectra API
  version: "1.0.0"
  description: |
    API publique Vectra — automatisez votre prospection B2B.

    ## Authentification
    Toutes les requêtes doivent inclure un header `Authorization: Bearer <your_key>`.

    - **Live** : clés `vt_live_xxx` — crédits consommés, données réelles.
    - **Sandbox** : clés `vt_test_xxx` — aucun crédit consommé, données mock.

    Générez vos clés depuis `/app/settings/developer`.

  contact:
    name: Support Vectra
    url: https://vectra-ai-saas-prospecting.vercel.app
servers:
  - url: https://vectra-ai-saas-prospecting.vercel.app/api/v1
    description: Production
  - url: http://localhost:3000/api/v1
    description: Local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "Format: Bearer vt_live_xxx (live) ou Bearer vt_test_xxx (sandbox)"
  schemas:
    Lead:
      type: object
      properties:
        id: { type: string, format: uuid }
        first_name: { type: string, nullable: true }
        last_name: { type: string, nullable: true }
        email: { type: string, format: email, nullable: true }
        company: { type: string, nullable: true }
        title: { type: string, nullable: true }
        linkedin_url: { type: string, nullable: true }
        score: { type: integer, minimum: 0, maximum: 100, nullable: true }
        status: { type: string, enum: [sourced, contacted, interested, not_interested, converted] }
        created_at: { type: string, format: date-time }
...
    WebhookEndpoint:
      type: object
      properties:
        id: { type: string, format: uuid }
        url: { type: string, format: uri }
        events: { type: array, items: { type: string } }
        is_active: { type: boolean }
        description: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
security:
  - bearerAuth: []
paths:
  /leads:
    get:
      tags: [Leads]
      summary: Lister les leads
      parameters:
        - in: query
          name: campaign_id
          schema: { type: string, format: uuid }
        - in: query
          name: q
          description: Recherche par nom, email ou entreprise
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 50, maximum: 200 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: Liste des leads
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Lead" } }
                  total: { type: integer }
                  limit: { type: integer }
                  offset: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
    post:
      tags: [Leads]
      summary: Créer un ou plusieurs leads (bulk)
      description: Envoyez un objet pour un lead, ou un tableau pour jusqu'à 500 leads.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/Lead"
                - type: array
                  items: { $ref: "#/components/schemas/Lead" }
                  maxItems: 500
      responses:
        "201":
          description: Leads créés
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Lead" } }
                  count: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
...
  /webhooks:
    get:
      tags: [Webhooks]
      summary: Lister les webhooks
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/WebhookEndpoint" } }
    post:
      tags: [Webhooks]
      summary: Créer un webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, events]
              properties:
                url: { type: string, format: uri, description: "URL HTTPS de réception" }
                events:
                  type: array
                  items:
                    type: string
                    enum:
                      - lead.created
                      - lead.updated
                      - lead.deleted
                      - message.generated
                      - message.approved
                      - message.sent
                      - sequence.step_sent
                      - sequence.completed
                      - sequence.stopped
                      - inbox.reply_received
                      - inbox.sentiment_changed
                description:
                  type: string
      responses:
        "201":
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/WebhookEndpoint" }
    delete:
      tags: [Webhooks]
      summary: Supprimer un webhook
      parameters:
        - in: query
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted: { type: boolean }
components:
  responses:
    Unauthorized:
      description: Clé API manquante ou invalide
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    RateLimited:
      description: Limite de requêtes atteinte
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/Error"
              - type: object
                properties:
                  retry_after: { type: integer, description: "Secondes avant de réessayer" }
...
    UnprocessableEntity:
      description: Données invalides
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }